+++ CHANGELOG.md
@@ -88,6 +88,7 @@ CHANGE LOG
 * Transforms were being hard-coded to include their experimental browser prefixes. These now use SC.browser.experimentalFooNameFor to correctly detect when browser prefixes aren't needed (or supported).
 * A great deal of dev-mode-only validation of layout properties has been added.
 * Separates SC.View#isFixedSize property into isFixedWidth and isFixedHeight, allowing for an even finer-grained optimization of child view resize notifications.
+* Adds a Developer Error when setting up a nested record attribute with an inverse property (via toOne or toMany).
 
 ### DEPRECATIONS & REMOVALS